+++ src/middlewares/validate.ts
@@ -10,7 +10,7 @@ export const validateQuery =
   <T>(schema: z.Schema<T>): RequestHandler =>
   (req: Request, res: Response, next: NextFunction) => {
     try {
-      req.query = schema.parse(req.query) as any;
+      req.query = schema.parse(req.query) as Record<string, any>;
       next();
     } catch (error) {
       if (error instanceof z.ZodError)
